+++ frontend_tests/node_tests/markdown_parse.js
@@ -52,7 +52,7 @@ function get_user_group_from_name(name) {
     return user_group_map.get(name);
 }
 
-function is_member_of_user_group(user_group_id, user_id) {
+function is_member_of_user_group(user_id, user_group_id) {
     assert.equal(user_group_id, staff_group.id);
     assert.equal(user_id, my_id);
     return true;