+++ frontend_tests/node_tests/user_groups.js
@@ -15,6 +15,7 @@ run_test("user_groups", () => {
         id: 0,
         members: new Set([1, 2]),
         is_system_group: false,
+        subgroups: new Set([4, 5]),
     };
 
     const params = {};
@@ -31,12 +32,14 @@ run_test("user_groups", () => {
         id: 1,
         members: new Set([3]),
         is_system_group: false,
+        subgroups: new Set([]),
     };
     const all = {
         name: "Everyone",
         id: 2,
         members: new Set([1, 2, 3]),
         is_system_group: false,
+        subgroups: new Set([4, 5, 6]),
     };
 
     user_groups.add(admins);
@@ -87,14 +90,23 @@ run_test("user_groups", () => {
     const groups_of_users_nomatch = user_groups.get_user_groups_of_user(user_id_not_in_any_group);
     assert.equal(groups_of_users_nomatch.length, 0);
 
-    assert.ok(!user_groups.is_member_of(admins.id, 4));
-    assert.ok(user_groups.is_member_of(admins.id, 3));
+    assert.ok(!user_groups.is_direct_member_of(4, admins.id));
+    assert.ok(user_groups.is_direct_member_of(3, admins.id));
 
     user_groups.add_members(all.id, [5, 4]);
     assert.deepEqual(user_groups.get_user_group_from_id(all.id).members, new Set([1, 2, 3, 5, 4]));
 
     user_groups.remove_members(all.id, [1, 4]);
     assert.deepEqual(user_groups.get_user_group_from_id(all.id).members, new Set([2, 3, 5]));
+
+    user_groups.add_subgroups(all.id, [2, 3]);
+    assert.deepEqual(
+        user_groups.get_user_group_from_id(all.id).subgroups,
+        new Set([2, 3, 5, 4, 6]),
+    );
+
+    user_groups.remove_subgroups(all.id, [2, 4]);
+    assert.deepEqual(user_groups.get_user_group_from_id(all.id).subgroups, new Set([3, 5, 6]));
 
     assert.ok(user_groups.is_user_group(admins));
     const object = {
@@ -107,9 +119,125 @@ run_test("user_groups", () => {
     assert.equal(user_groups.get_realm_user_groups().length, 0);
 
     blueslip.expect("error", "Could not find user group with ID -1");
-    assert.equal(user_groups.is_member_of(-1, 15), false);
+    assert.equal(user_groups.is_direct_member_of(15, -1), false);
 
-    blueslip.expect("error", "Could not find user group with ID -9999", 2);
+    blueslip.expect("error", "Could not find user group with ID -9999", 4);
     user_groups.add_members(-9999);
     user_groups.remove_members(-9999);
+    user_groups.add_subgroups(-9999);
+    user_groups.remove_subgroups(-9999);
+});
+
+run_test("get_recursive_subgroups", () => {
+    const admins = {
+        name: "Admins",
+        description: "foo",
+        id: 1,
+        members: new Set([1]),
+        is_system_group: false,
+        subgroups: new Set([4]),
+    };
+    const all = {
+        name: "Everyone",
+        id: 2,
+        members: new Set([2, 3]),
+        is_system_group: false,
+        subgroups: new Set([1, 3]),
+    };
+    const test = {
+        name: "Test",
+        id: 3,
+        members: new Set([3, 4, 5]),
+        is_system_group: false,
+        subgroups: new Set([2]),
+    };
+    const foo = {
+        name: "Foo",
+        id: 4,
+        members: new Set([6, 7]),
+        is_system_group: false,
+        subgroups: new Set([]),
+    };
+
+    user_groups.add(admins);
+    user_groups.add(all);
+    user_groups.add(test);
+    user_groups.add(foo);
+
+    // This test setup has a state that won't appear in real data: Groups 2 and 3
+    // each contain the other. We test this corner case because it is a simple way
+    // to verify whether our algorithm correctly avoids visiting groups multiple times
+    // when determining recursive subgroups.
+    // A test case that can occur in practice and would be problematic without this
+    // optimization is a tree where each layer connects to every node in the next layer.
+    assert.deepEqual(user_groups.get_recursive_subgroups(admins.id), new Set([4]));
+    assert.deepEqual(user_groups.get_recursive_subgroups(all.id), new Set([4, 1, 2, 3]));
+    assert.deepEqual(user_groups.get_recursive_subgroups(test.id), new Set([2, 4, 1, 3]));
+    assert.deepEqual(user_groups.get_recursive_subgroups(foo.id), new Set([]));
+
+    blueslip.expect("error", "Could not find user group with ID 1111");
+    assert.deepEqual(user_groups.get_recursive_subgroups(1111), undefined);
+
+    user_groups.add_subgroups(foo.id, [9999]);
+    blueslip.expect("error", "Could not find subgroup with ID 9999", 2);
+    assert.deepEqual(user_groups.get_recursive_subgroups(foo.id), undefined);
+    assert.deepEqual(user_groups.get_recursive_subgroups(test.id), undefined);
+});
+
+run_test("is_user_in_group", () => {
+    const admins = {
+        name: "Admins",
+        id: 1,
+        members: new Set([1]),
+        is_system_group: false,
+        subgroups: new Set([4]),
+    };
+    const all = {
+        name: "Everyone",
+        id: 2,
+        members: new Set([2, 3]),
+        is_system_group: false,
+        subgroups: new Set([1, 3]),
+    };
+    const test = {
+        name: "Test",
+        id: 3,
+        members: new Set([4, 5]),
+        is_system_group: false,
+        subgroups: new Set([1]),
+    };
+    const foo = {
+        name: "Foo",
+        id: 4,
+        members: new Set([6, 7]),
+        is_system_group: false,
+        subgroups: new Set([]),
+    };
+    user_groups.add(admins);
+    user_groups.add(all);
+    user_groups.add(test);
+    user_groups.add(foo);
+
+    assert.equal(user_groups.is_user_in_group(admins.id, 1), true);
+    assert.equal(user_groups.is_user_in_group(admins.id, 6), true);
+    assert.equal(user_groups.is_user_in_group(admins.id, 3), false);
+
+    assert.equal(user_groups.is_user_in_group(all.id, 2), true);
+    assert.equal(user_groups.is_user_in_group(all.id, 1), true);
+    assert.equal(user_groups.is_user_in_group(all.id, 6), true);
+
+    assert.equal(user_groups.is_user_in_group(test.id, 4), true);
+    assert.equal(user_groups.is_user_in_group(test.id, 1), true);
+    assert.equal(user_groups.is_user_in_group(test.id, 6), true);
+    assert.equal(user_groups.is_user_in_group(test.id, 3), false);
+
+    assert.equal(user_groups.is_user_in_group(foo.id, 6), true);
+    assert.equal(user_groups.is_user_in_group(foo.id, 3), false);
+
+    blueslip.expect("error", "Could not find user group with ID 1111");
+    assert.equal(user_groups.is_user_in_group(1111, 3), false);
+
+    user_groups.add_subgroups(foo.id, [9999]);
+    blueslip.expect("error", "Could not find subgroup with ID 9999");
+    assert.equal(user_groups.is_user_in_group(admins.id, 6), false);
 });