+++ static/js/popovers.js
@@ -42,6 +42,7 @@ import * as realm_playground from "./realm_playground";
 import * as reminder from "./reminder";
 import * as resize from "./resize";
 import * as rows from "./rows";
+import * as settings_config from "./settings_config";
 import * as settings_data from "./settings_data";
 import * as settings_users from "./settings_users";
 import * as stream_popover from "./stream_popover";
@@ -205,6 +206,7 @@ function render_user_info_popover(
     }
 
     const muting_allowed = !is_me && !user.is_bot;
+    const is_active = people.is_active_user_for_popover(user.user_id);
     const is_muted = muted_users.is_user_muted(user.user_id);
     const status_text = user_status.get_status_text(user.user_id);
     const status_emoji_info = user_status.get_status_emoji(user.user_id);
@@ -221,9 +223,14 @@ function render_user_info_popover(
         can_set_away,
         can_mute: muting_allowed && !is_muted,
         can_manage_user: page_params.is_admin && !is_me,
+        can_send_private_message:
+            is_active &&
+            !is_me &&
+            page_params.realm_private_message_policy !==
+                settings_config.private_message_policy_values.disabled.code,
         can_unmute: muting_allowed && is_muted,
         has_message_context,
-        is_active: people.is_active_user_for_popover(user.user_id),
+        is_active,
         is_bot: user.is_bot,
         is_me,
         is_sender_popover,
@@ -716,6 +723,12 @@ export function hide_user_sidebar_popover() {
     }
 }
 
+function hide_all_user_info_popovers() {
+    hide_message_info_popover();
+    hide_user_sidebar_popover();
+    hide_user_info_popover();
+}
+
 function focus_user_info_popover_item() {
     // For now I recommend only calling this when the user opens the menu with a hotkey.
     // Our popup menus act kind of funny when you mix keyboard and mouse.
@@ -1005,9 +1018,7 @@ export function register_click_handlers() {
 
     $("body").on("click", ".info_popover_actions .sidebar-popover-mute-user", (e) => {
         const user_id = elem_to_user_id($(e.target).parents("ul"));
-        hide_message_info_popover();
-        hide_user_sidebar_popover();
-        hide_user_info_popover();
+        hide_all_user_info_popovers();
         e.stopPropagation();
         e.preventDefault();
         muted_users_ui.confirm_mute_user(user_id);
@@ -1015,9 +1026,7 @@ export function register_click_handlers() {
 
     $("body").on("click", ".info_popover_actions .sidebar-popover-unmute-user", (e) => {
         const user_id = elem_to_user_id($(e.target).parents("ul"));
-        hide_message_info_popover();
-        hide_user_sidebar_popover();
-        hide_user_info_popover();
+        hide_all_user_info_popovers();
         muted_users_ui.unmute_user(user_id);
         e.stopPropagation();
         e.preventDefault();
@@ -1316,7 +1325,6 @@ export function hide_all_except_sidebars(opts) {
         hideAll({exclude: opts.exclude_tippy_instance});
     }
     hide_actions_popover();
-    hide_message_info_popover();
     emoji_picker.hide_emoji_popover();
     giphy.hide_giphy_popover();
     stream_popover.hide_stream_popover();
@@ -1324,8 +1332,7 @@ export function hide_all_except_sidebars(opts) {
     stream_popover.hide_all_messages_popover();
     stream_popover.hide_starred_messages_popover();
     stream_popover.hide_drafts_popover();
-    hide_user_sidebar_popover();
-    hide_user_info_popover();
+    hide_all_user_info_popovers();
     hide_playground_links_popover();
 
     // look through all the popovers that have been added and removed.