+++ zerver/tests/test_docs.py
@@ -214,6 +214,7 @@ class DocPageTest(ZulipTestCase):
         self._test("/case-studies/tum/", "Technical University of Munich")
         self._test("/case-studies/ucsd/", "UCSD")
         self._test("/case-studies/rust/", "Rust programming language")
+        self._test("/case-studies/recurse-center/", "Recurse Center")
         self._test("/case-studies/lean/", "Lean theorem prover")
         self._test("/case-studies/idrift/", "Case study: iDrift AS")
         self._test("/case-studies/asciidoctor/", "Case study: Asciidoctor")
@@ -463,7 +464,7 @@ class PlansPageTest(ZulipTestCase):
         result = self.client_get("/plans/", subdomain=root_domain)
         self.assert_in_success_response(["Self-host Zulip"], result)
         self.assert_not_in_success_response(["/upgrade#sponsorship"], result)
-        self.assert_in_success_response(["/accounts/go/?next=/upgrade%23sponsorship"], result)
+        self.assert_in_success_response(["/accounts/go/?next=%2Fupgrade%23sponsorship"], result)
 
         non_existent_domain = "moo"
         result = self.client_get("/plans/", subdomain=non_existent_domain)
@@ -487,7 +488,7 @@ class PlansPageTest(ZulipTestCase):
         result = self.client_get("/plans/", subdomain="zulip")
         self.assert_in_success_response(["Current plan"], result)
         self.assert_in_success_response(["/upgrade#sponsorship"], result)
-        self.assert_not_in_success_response(["/accounts/go/?next=/upgrade%23sponsorship"], result)
+        self.assert_not_in_success_response(["/accounts/go/?next=%2Fupgrade%23sponsorship"], result)
 
         # Test root domain, with login on different domain
         result = self.client_get("/plans/", subdomain="")